+++ libretro-common/formats/libchdr/libchdr_lzma.c
@@ -148,7 +148,7 @@ void *lzma_fast_alloc(void *p, size_t size)
 	}
 
 	/* set the low bit of the size so we don't match next time */
-	*addr = size | 1;
+	*addr = ((uint32_t)size) | 1;
 
 	/* return aligned address */
 	return (void*)vaddr;
